+++ ios/Classes/Extension/Models/PaymentChannel+Compare.swift
@@ -1,0 +1,7 @@
+import Tpay
+
+extension Headless.Models.PaymentChannel {
+    func has(channelId: Int) -> Bool {
+        return String(channelId) == self.id
+    }
+}