+++ ios/Classes/Model/Domain/BankPayment.swift
@@ -10,4 +10,5 @@ struct BankPayment: Transaction {
     let payerContext: PayerContext?
     let paymentChannel: Headless.Models.PaymentChannel
     let callbacks: CallbacksConfiguration
+    let notification: Tpay.Notification?
 }