+++ ios/Classes/Model/Domain/BlikPayment.swift
@@ -12,4 +12,5 @@ struct BlikPayment: Transaction {
     let alias: String?
     let paymentChannel: Headless.Models.PaymentChannel
     let callbacks: CallbacksConfiguration
+    let notification: Tpay.Notification?
 }