+++ ios/Classes/Model/Domain/CardPayment.swift
@@ -12,4 +12,5 @@ struct CardPayment: Transaction {
     let card: Headless.Models.Card?
     let cardToken: Headless.Models.CardToken?
     let callbacks: CallbacksConfiguration
+    let notification: Tpay.Notification?
 }