+++ ios/Classes/Model/Domain/DigitalWalletPayment.swift
@@ -11,4 +11,5 @@ struct DigitalWalletPayment: Transaction {
     let paymentChannel: Headless.Models.PaymentChannel
     let token: String
     let callbacks: CallbacksConfiguration
+    let notification: Tpay.Notification?
 }