+++ lib/model/screenless/credit_card_payment.g.dart
@@ -11,9 +11,7 @@ CreditCardPayment _$CreditCardPaymentFromJson(Map<String, dynamic> json) =>
       paymentDetails: PaymentDetails.fromJson(
           json['paymentDetails'] as Map<String, dynamic>),
       payer: Payer.fromJson(json['payer'] as Map<String, dynamic>),
-      callbacks: json['callbacks'] == null
-          ? null
-          : Callbacks.fromJson(json['callbacks'] as Map<String, dynamic>),
+      callbacks: Callbacks.fromJson(json['callbacks'] as Map<String, dynamic>),
       creditCard: json['creditCard'] == null
           ? null
           : CreditCard.fromJson(json['creditCard'] as Map<String, dynamic>),
@@ -27,7 +25,7 @@ Map<String, dynamic> _$CreditCardPaymentToJson(CreditCardPayment instance) =>
     <String, dynamic>{
       'paymentDetails': instance.paymentDetails.toJson(),
       'payer': instance.payer.toJson(),
-      'callbacks': instance.callbacks?.toJson(),
+      'callbacks': instance.callbacks.toJson(),
       'creditCard': instance.creditCard?.toJson(),
       'creditCardToken': instance.creditCardToken,
       'recursive': instance.recursive?.toJson(),