+++ lib/model/screenless/pay_po_payment.g.dart
@@ -10,9 +10,7 @@ PayPoPayment _$PayPoPaymentFromJson(Map<String, dynamic> json) => PayPoPayment(
       paymentDetails: PaymentDetails.fromJson(
           json['paymentDetails'] as Map<String, dynamic>),
       payer: Payer.fromJson(json['payer'] as Map<String, dynamic>),
-      callbacks: json['callbacks'] == null
-          ? null
-          : Callbacks.fromJson(json['callbacks'] as Map<String, dynamic>),
+      callbacks: Callbacks.fromJson(json['callbacks'] as Map<String, dynamic>),
     );
 
 Map<String, dynamic> _$PayPoPaymentToJson(PayPoPayment instance) =>