+++ lib/model/screenless/raty_pekao_payment.g.dart
@@ -11,9 +11,7 @@ RatyPekaoPayment _$RatyPekaoPaymentFromJson(Map<String, dynamic> json) =>
       paymentDetails: PaymentDetails.fromJson(
           json['paymentDetails'] as Map<String, dynamic>),
       payer: Payer.fromJson(json['payer'] as Map<String, dynamic>),
-      callbacks: json['callbacks'] == null
-          ? null
-          : Callbacks.fromJson(json['callbacks'] as Map<String, dynamic>),
+      callbacks: Callbacks.fromJson(json['callbacks'] as Map<String, dynamic>),
       channelId: (json['channelId'] as num).toInt(),
     );