+++ lib/model/screenless/screenless_payment.g.dart
@@ -11,7 +11,5 @@ ScreenlessPayment _$ScreenlessPaymentFromJson(Map<String, dynamic> json) =>
       paymentDetails: PaymentDetails.fromJson(
           json['paymentDetails'] as Map<String, dynamic>),
       payer: Payer.fromJson(json['payer'] as Map<String, dynamic>),
-      callbacks: json['callbacks'] == null
-          ? null
-          : Callbacks.fromJson(json['callbacks'] as Map<String, dynamic>),
+      callbacks: Callbacks.fromJson(json['callbacks'] as Map<String, dynamic>),
     );