+++ lib/model/screenless/transfer_payment.g.dart
@@ -11,11 +11,8 @@ TransferPayment _$TransferPaymentFromJson(Map<String, dynamic> json) =>
       paymentDetails: PaymentDetails.fromJson(
           json['paymentDetails'] as Map<String, dynamic>),
       payer: Payer.fromJson(json['payer'] as Map<String, dynamic>),
-      callbacks: json['callbacks'] == null
-          ? null
-          : Callbacks.fromJson(json['callbacks'] as Map<String, dynamic>),
+      callbacks: Callbacks.fromJson(json['callbacks'] as Map<String, dynamic>),
       channelId: (json['channelId'] as num).toInt(),
-      bankName: json['bankName'] as String,
     );
 
 Map<String, dynamic> _$TransferPaymentToJson(TransferPayment instance) =>
@@ -24,5 +21,4 @@ Map<String, dynamic> _$TransferPaymentToJson(TransferPayment instance) =>
       'payer': instance.payer,
       'callbacks': instance.callbacks,
       'channelId': instance.channelId,
-      'bankName': instance.bankName,
     };