+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## UNRELEASED
+
+## 1.2.7
+- Fix: Updated iOS SDK to version 1.3.6 to support exact channel in screenless mode for `TransferPayment`
+- Change: Updated iOS SDK to version 1.3.6 to handle passing `notifications` parameter
+- Fix: Screenless payments parsing on iOS
+- Change: Screenless payments now ___REQUIRES___ `callbacks` to be provided
+
 ## 1.2.6
 - Updated README.md file with improved documentation